API

Base.read — Method

read(tsg::TasmanianSG, filename)

reads the tsg grid from a file discards any existing grid held by tsg

tsg: an existing grid filename: string indicating a grid file where a grid was already written using write from Julia or any other Tasmanian interface

output: Bool true: the read was successful false: the read failed, check for an error message

Base.write — Method

write(tsg::TasmanianSG, filename; binary::Bool=true)

writes the tsg grid to a file

tsg: an existing grid filename: string indicating a grid file where a grid will be written binary: Bool true: write to a binary file false: write to an ASCII file

Tasmanian.clearConformalTransform — Method

clearConformalTransform(tsg::TasmanianSG)

resets the conformal domain transform loaded values will be kept, however, the values now correspond to canonical points and may be invalid for your application

Tasmanian.clearDomainTransform — Method

clearDomainTransform(tsg::TasmanianSG)

resets the domain to canonical loaded values will be kept, however, the values now correspond to canonical points and may be invalid for your application

Tasmanian.clearLevelLimits — Method

clearLevelLimits(tsg::TasmanianSG)

clears the limits set by the last make***Grid or refine command if no limits are set, this has no effect

Tasmanian.clearRefinement — Method

clearRefinement(tsg::TasmanianSG)

clear the last call to set***Refinement, only works if called before the points are loaded, i.e., before loadNeededPoints()

if getNumNeeded() == 0, this call will have no effect

Tasmanian.copyGrid — Function

copyGrid(tsg::TasmanianSG, outputs_begin = 0, outputs_end = -1)

accepts an instance of TasmanianSparseGrid class and creates a hard copy of the class and all included data original class is not modified

tsg: instance of TasmanianSG the source for the copy

outputs_begin: integer indicating the first output to copy

outputsend: integer one bigger than the last output to copy if set to -1, all outputs from outputsbegin to the end will be copied

Examples:

copyGrid(other, 0, -1) # copy all outputs (default) copyGrid(other, 0, getNumOutputs(other)) # also copy all copyGrid(other, 0, 3) # copy outputs 0, 1, and 2 copyGrid(other, 1, 4) # copy outputs 1, 2, and 3

Tasmanian.differentiate! — Method

differentiate!(jacobian::VecOrMat{Float64}, tsg::TasmanianSG, x::AbstractVector{Float64})

returns the derivative (Jacobian or gradient vector) of the interpolant

jacobian: a vector or a matrix with dimensions dimensions X outputs each row corresponds to the value of the interpolant for one columns of vals tsg: an instance of TasmanianSG x: a vector with length dimensions which is the evaluation point

Tasmanian.differentiate — Method

differentiate(tsg::TasmanianSG, x::Vector{Float64})

returns the derivative (Jacobian or gradient vector) of the interpolant as a vector or a matrix with dimensions dimensions X outputs each row corresponds to the value of the interpolant

tsg: an instance of TasmanianSG x: a vector with length iDimensions which is the evaluation point

Tasmanian.enableAcceleration — Method

enableAcceleration(tsg::TasmanianSG, acceleration_type, GPUID = 0)

Enables the use of accelerated backend libraries and extensions, such as BLAS and CUDA. Each acceleration type requires corresponding CMake compile options, otherwise the backend will fallback to the closest available options.

sAccelerationType: string

'none' core fallback mode, relies on sequential implementation if compiled with TasmanianENABLEOPENMP this will use simple "omp parallel for" to take advantage of multiple cpu cores

'cpu-blas' uses BLAS level 2 and 3 functions for acceleration of batch evaluations requires TasmanianENABLEBLAS=ON this is the default mode, if available

'gpu-default' uses CUDA kernels, cuBlas, cuSparse and MAGMA libraries for accelerated matrix operations, e.g., cublasDgemm refer to TasGrid::TypeAcceleration for more details

'gpu_cublas' uses the Nvidia cuBlas and cuSparse libraries

'gpu-cuda' uses custom CUDA kernels in addition to the accelerated linear algebra libraries

'gpu-magma' uses the custom CUDA kernels and the MAGMA library in place of the default Nvidia libraries

GPUID: integer indicates the GPU device to use, if set to None then device zero will be used first or the device set with setGPUID()

Tasmanian.estimateAnisotropicCoefficients — Method

estimateAnisotropicCoefficients(tsg::TasmanianSG, type, output)

returns the estimate of the anisotropic coefficients from the current set of loaded points see the manual

type: string identifying the estimate to use (see the Manual) recommended: 'iptotal' 'ipcurved'

output: int (indicates the output to use) selects which output to use for refinement sequence grids accept -1 to indicate all outputs

returns vector of length getNumDimensions() or 2*getNumDimensions() the first set of getNumDimensions() entries correspond to the xi coefficients the second set of getNumDimensions() entries correspond to the eta coefficients

Tasmanian.evaluate — Method

evaluate(tsg::TasmanianSG, x)

evaluates the intepolant at a single points of interest and returns the result This is the accelerated version using the selected acceleration type, but it is potentially not thread safe

this should be called after the grid has been created and after values have been loaded

x: a vector with length getNumDimensions() the entries indicate the points for evaluating the weights

output: returns vector of length getNumOutputs() the values of the interpolant at fX

Tasmanian.evaluateBatch! — Method

 evaluateBatch!(y::AbstractVecOrMat{Float64}, tsg::TasmanianSG, vals::AbstractVecOrMat{Float64})

evaluates the intepolant at the points of interest and set the result in y

this should be called after the grid has been created and after values have been loaded

y: a vector or a matrix with dimensions outputs X size(vals, 2) each columns corresponds to the value of the interpolant for one columns of vals

vals: a vector or a matrix with first dimension equal to dimensions each column in the array is a single requested point

Tasmanian.evaluateBatch — Method

 evaluateBatch(tsg::TasmanianSG, vals::AbstractVecOrMat{Float64})

evaluates the intepolant at the points of interest and returns the result

this should be called after the grid has been created and after values have been loaded

vals: a vector or a matrix with first dimension equal to dimensions each column in the array is a single requested point

output: a vector or a matrix with dimensions outputs X size(vals, 2) each columns corresponds to the value of the interpolant for one columns of vals

Tasmanian.evaluateHierarchicalFunctions — Method

evaluateHierarchicalFunctions(tsg::TasmanianSG, x)

evaluates the hierarchical functions at a set of points in the domain and return a matrix with the result

x: a matrix of dimensions tsg.dimensions by getNumPoints(tsg) the columns indicate the points for evaluating the weights

output: returns a matrix of dimensions shape == [x.shape[0], getNumPoints()] the values of the basis functions at the points

Tasmanian.evaluateSparseHierarchicalFunctions — Method

evaluateSparseHierarchicalFunctions(tsg::TasmanianSG, x)

evaluates the hierarchical functions at a set of points in the domain. The distinction between this function and evaluateHierarchicalFunctions() lies in the type of the returned result, namely a sparse vs a dense matrix.

The motivation for this function is that Local Polynomial and Wavelet grids usually result in sparse matrices

x: a matrix with dimensions getNumDimensions() by NumX the entries indicate the points for evaluating

output: returns a TasmanianSimpleSparseMatrix class which is a simple class with three fields: aPntr, aIndx, and aVals which are numpy.ndarray of types int32, int32, and float64 NumRows and NumCols are meta fields and have values NumRows = x.shape[0] NumCols = getNumPoints() The sparse matrix is compressed along the x.shape[0] dimension, i.e., using column compressed format

Tasmanian.evaluateThreadSafe — Method

evaluateThreadSafe(tsg::TasmanianSG, x) evaluates the intepolant at a single points of interest and returns the result This is the thread safe version, but it does not use acceleration of any type

    this should be called after the grid has been created and after
    values have been loaded

    x: vector with length dimensions
       the entries indicate the points for evaluating the weights

    output: returns a vector of length getNumOutputs
            the values of the interpolant at x

Tasmanian.getAccelerationType — Method

getAccelerationType(tsg::TasmanianSG)

returns the type of acceleration set by enableAcceleration

Tasmanian.getAlpha — Method

getAlpha(tsg::TasmanianSG)

returns the value of alpha in the call to makeGlobalGrid
if makeGlobalGrid has not been called, returns 0.0

Tasmanian.getAnisotropicRefinement — Function

getAnisotropicRefinement(tsg::TasmanianSG, type, min_growth, output, level_Limits = Vector{Int32}(undef, 0))

Calls setAnistropicRefinement() on the inputs and then getNeededPoints().

Tasmanian.getBeta — Method

getBeta(tsg::TasmanianSG)

returns the value of beta in the call to makeGlobalGrid
if makeGlobalGrid has not been called, returns 0.0

Tasmanian.getConformalTransformASIN — Method

getConformalTransformASIN(tsg::TasmanianSG)

returns Truncation from the call to setConformalTransformASIN()

if setConformalTransformASIN() has not been called or if the transformed has been cleared by clearConformalTransform(), then this returns an empty matrix

Tasmanian.getDomainTransform — Method

getDomainTransform(tsg::TasmanianSG)

returns Transform from the call to setDomainTransform()

if setDomainTransform() has not been called or if the transformed has been cleared by clearDomainTransform(), then this returns an empty matrix

Tasmanian.getGPUID — Method

getGPUID(tsg::TasmanianSG)

returns the GPU ID set using setGPUID

Tasmanian.getGPUMemory — Method

getGPUMemory(tsg::TasmanianSG, GPUID)

returns the total memory (in MegaBytes, 1024**2 bytes) of the corresponding GPU

this is one of several functions designed to allow basic management of multi-gpu setup with only Tasmanian module

Tasmanian.getGPUName — Method

getGPUName(tsg::TasmanianSG, GPUID)

return the cuda name ID of the corresponding GPU

this is one of several functions designed to allow basic management of multi-gpu setup with only Tasmanian module

Tasmanian.getGlobalPolynomialSpace — Method

getGlobalPolynomialSpace(tsg::TasmanianSG, interpolation)

returns a matrix corresponding to the polynomial space that is integrated or interpolated exactly by the current grid

interpolation: boolean indicates whether to give the space associated with integration or interpolation

output: is a matrix of integers size(output, 1) is equal to iDimension size(output, 2) indicates the cardinality of the space each columns corresponds to a multi-index associated with a polynomial in a hierarchical tensor basis, for example, monomials

    see the manual for details

Tasmanian.getHierarchicalCoefficients — Method

getHierarchicalCoefficients(tsg::TasmanianSG)

For global grids, this just returns the values loaded using the call to loadNeededPoints(). In all other cases, this returns the list of hierarchical coefficients, i.e., surpluses.

returns a matrix getNumOutputs() by getNumPoints()

Tasmanian.getHierarchicalSupport — Method

getHierarchicalSupport(tsg::TasmanianSG)

returns the support of the hierarchical basis in a matrix

only local-polynomial and wavelet grids have restricted support

the support of all basis is restricted to the domain even if the support includes additional ares

output: a matrix of dimensions getNumDimensions() by getNumPoints() the support entries correspond to the output of getPoints() if x represents a point in the domain and if abs( x[j] - getPoints()[j, i] ) > getSupport()[j, i] then the i-th entry in evaluateHierarchicalFunctions(x) corresponding to x is guaranteed to be zero

Tasmanian.getInterpolationWeights — Method

getInterpolationWeights(tsg, x)

returns the interpolation weights associated with the points in getPoints()

x: a vector with length dimensions the entries indicate the points for evaluating the weights

output: a vector of length getNumPoints() the order of the weights matches the order in getPoints()

Tasmanian.getInterpolationWeightsBatch — Method

getInterpolationWeightsBatch(tsg, llfX):

returns the interpolation weights associated with the points in getPoints()

finds multiple weights with a single library call uses OpenMP if enabled in libtasmaniansparsegrids.so

x: a matrix with first dimension getNumDimensions() each column in the array is a single requested point

output: a matrix with dimensions getNumPoints() X size(llfX, 2) each row corresponds to the weight for one row of llfX

Tasmanian.getLevelLimits — Method

getLevelLimits(tsg::TasmanianSG)

returns the limits set by the last call to make***Grid or refine returns a vector of integers corresponding to the limits for each direction, -1 indicates no limit

Tasmanian.getLicense — Method

getLicense()

returns Tasmanian library hardcoded license string

Tasmanian.getLoadedPoints — Method

getLoadedPoints(tsg::TasmanianSG)

returns the points loaded in the existing interpolant

output: matrix of size getNumDimensions() X getNumNeeded() each column corresponds to one point if getNumLoaded() == 0, returns zeros(2)

Tasmanian.getLoadedValues — Method

getLoadedValues(tsg::TasmanianSG)

Returns the model values as given to Tasmanian by the loadNeededPoints() method. The ordering will match the current internal ordering, e.g., mixing the different model values from different refinement iterations.

Returns a matrix with size getNumOutputs() by getNumPoints()

Tasmanian.getNeededPoints — Method

getNeededPoints(tsg::TasmanianSG)

returns the points needed to form the interpolant or the next level of refinement following a set***Refinement() call

output: 2-D array of size dimension X getNumNeeded() each column corresponds to one point if (getNumNeeded() == 0): returns zeros(dimensions, 0)

Tasmanian.getNumDimensions — Method

getNumDimensions(tsg::TasmanianSG)

returns the value of iDimension in the make***Grid command if no grid has been made, it returns 0

Tasmanian.getNumGPUs — Method

getNumGPUs()
    returns the number of available GPUs according to cuda

    this is one of several functions designed to allow basic
    management of multi-gpu setup with only Tasmanian module

Tasmanian.getNumLoaded — Method

getNumLoaded(tsg::TasmanianSG)

returns the number of points loaded in the existing interpolant

Tasmanian.getNumNeeded — Method

getNumNeeded(tsg::TasmanianSG)

returns the number of points needed to form the interpolant or form the next interpolant following a refinement

Tasmanian.getNumOutputs — Method

getNumOutputs(tsg::TasmanianSG)

returns the value of iOutputs in the make***Grid command if no grid has been made, it returns 0

Tasmanian.getNumPoints — Method

getNumPoints(tsg::TasmanianSG)

if points have been loaded, returns the same as getNumLoaded() otherwise, returns the same as getNumNeeded()

Tasmanian.getOrder — Method

getOrder(tsg::TasmanianSG)

returns the value of iOrder in the call to makeLocalPolynomialGrid or makeWaveletGrid if makeLocalPolynomialGrid and makeWaveletGrid have not been called, returns -1

Tasmanian.getPoints — Method

getPoints(tsg::TasmanianSG)

if points have been loaded, gives the same as getLoadedPoints() otherwise, returns the same as getNeededPoints()

Tasmanian.getQuadratureWeights — Method

getQuadratureWeights(tsg)

returns the quadrature weights associated with the points in getPoints()

output: a vector of length getNumPoints() the order of the weights matches the order in getPoints()

Tasmanian.getRule — Method

getRule(tsg::TasmanianSG)

returns the value of rule in the make***Grid command if makeWaveletGrid is used, returns "wavelet" if no grid has been made, it returns "unknown"

Tasmanian.getSurplusRefinement — Function

getSurplusRefinement(tsg::TasmanianSG, tolerance, output, criteria = "", level_limits = Vector{Int32}(undef, 0), scale_correction = [])

Calls setSurplusRefinement() on the inputs and then getNeededPoints().

Tasmanian.getVersion — Method

getVersion()

returns Tasmanian library hardcoded version

Tasmanian.get_VersionMajor — Method

get_VersionMajor()

returns the hardcoded version major int

Tasmanian.get_VersionMinor — Method

get_VersionMinor()

returns the hardcoded version minor int

Tasmanian.integrate — Method

integrate(tsg::TasmanianSG)

returns the integral of the interpolant

output: returns a vector of length outputs the integral of the interpolant

Tasmanian.integrateHierarchicalFunctions — Method

integrateHierarchicalFunctions(tsg::TasmanianSG)

Computes the integrals of the hierarchical basis functions, i.e., the same functions computed by evaluateHierarchicalFunctions().

returns a vector of length getNumPoints()

Tasmanian.isAccelerationAvailable — Method

isAccelerationAvailable(tsg::TasmanianSG, acceleration_type)

returns true if the library has been compiled with support for acceleration_type. Even if this returns false, you can use the type for enableAcceleration, but the library will default to the next available type (see the Manual)

Tasmanian.isCudaEnabled — Method

isCudaEnabled()

returns true if the library has been computed with Cuda support

Tasmanian.isDpcppEnabled — Method

isDpcppEnabled()

returns true if the library has been computed with DPC++ support

Tasmanian.isFourier — Method

isFourier(tsg::TasmanianSG)

returns true if using a Fourier grid

Tasmanian.isGlobal — Method

isGlobal(tsg::TasmanianSG)

returns true if using a global grid

Tasmanian.isHipEnabled — Method

isHipEnabled()

returns true if the library has been computed with HIP support

Tasmanian.isLocalPolynomial — Method

isLocalPolynomial(tsg::TasmanianSG)

returns true if using a local polynomial grid

Tasmanian.isOpenMPEnabled — Method

isOpenMPEnabled()

returns true if the library has been computed with OpenMP support

Tasmanian.isSequence — Method

isSequence(tsg::TasmanianSG)

returns true if using a sequence grid

Tasmanian.isSetConformalTransformASIN — Method

isSetConformalTransformASIN(tsg::TasmanianSG)

returns true if conformal transform is set returns false otherwise

see: setConformalTransformASIN()

Tasmanian.isSetDomainTransform — Method

isSetDomainTransform(tsg::TasmaninaSG) returns true if the grid is defined for non-canonical domain returns false if using a canonical domain

Tasmanian.isWavelet — Method

isWavelet(tsg::TasmanianSG)

returns true if using a local wavelet grid

Tasmanian.loadNeededPoints! — Method

loadNeededPoints!(tsg::TasmanianSG, vals::Array{Float64})

loads the values of the target function at the needed points if there are no needed points, this reset the currently loaded values

vals: an array with dimensions outputs X getNumNeeded() each column corresponds to the values of the outputs at the corresponding needed point. The order and leading dimension must match the points obtained from getNeededPoints()

Tasmanian.loadNeededValues! — Method

loadNeededValues!(tsg::TasmanianSG, vals::Array{Float64})

Alias of loadNeededPoints()

Tasmanian.makeFourierGrid! — Method

makeFourierGrid!(tsg::TasmanianSG; type, anisotropic_weights=Vector{Int32}(undef, 0), level_limits=Vector{Int32}(undef, 0))

creates a new sparse grid using a Fourier rule discards any existing grid held by this class

type: string identifying the tensor selection strategy 'level' 'curved' 'hyperbolic' 'tensor' 'iptotal' 'ipcurved' 'iphyperbolic' 'iptensor' 'qptotal' 'qpcurved' 'qphyperbolic' 'qptensor'

anisotropic_weights: list or array of weights length must be iDimension or 2*iDimension the first tsg.dimensions weights must be positive see the manual for details

Tasmanian.makeGlobalGrid! — Method

makeGlobalGrid!(tsg::TasmanianSG; type, rule, anisotropic_weights=Vector{Int32}(undef, 0), alpha=0.0, beta=0.0, custom_filename="", level_limits=Vector{Int32}(undef, 0))

creates a new sparse grid using a global polynomial rule discards any existing grid held by tsg

type: string identifying the tensor selection strategy level curved hyperbolic tensor iptotal ipcurved iphyperbolic iptensor qptotal qpcurved qphyperbolic qptensor

rule: string (defines the 1-D rule that induces the grid)

   Interpolation rules

      Note: the quadrature induced by those rules is constructed
            by integrating the interpolant

      `clenshaw-curtis`    `clenshaw-curtis-zero`      `fejer2`
      `rleja`    `rleja-odd`  `rleja-double2`   `rleja-double4`
      `rleja-shifted`   `rleja-shifted-even`
      `max-lebesgue`    `max-lebesgue-odd`
      `min-lebesgue`    `min-lebesgue-odd`
      `leja`            `leja-odd`
      `min-delta`       `min-delta-odd`

      `chebyshev`       `chebyshev-odd`
        approximation using roots of Chebyshev polynomials
        non-nested case (in contrast to Clenshaw-Curtis nodes)

   Quadrature rules, the weights target exactness with respect
                     to the highest polynomial degree possible

       `gauss-legendre`  `gauss-legendre-odd`
        approximation using roots of polynomials orthogonal in
        measure Uniform

       `gauss-patterson`  (a.k.a. nested Gauss-Legendre)
        Note: the nodes and weights are hard-coded hence there
        is a limit on the highest possible depth

       `gauss-chebyshev1`  `gauss-chebyshev1-odd`
       `gauss-chebyshev2`  `gauss-chebyshev2-odd`
         approximation using roots of polynomials orthogonal in
         measures  1/sqrt(1-x^2) and sqrt(1-x^2)  (respectively)

      `gauss-gegenbauer`  `gauss-gegenbauer-odd`
        approximation using roots of polynomials orthogonal in
        measure (1-x^2)^alpha

      `gauss-jacobi`
        approximation using roots of polynomials orthogonal in
        measure (1-x)^alpha * (1+x)^beta

      `gauss-laguerre`
        approximation using roots of polynomials orthogonal in
        measure x^alpha * epx(-x)

      `gauss-hermite`  `gauss-hermite-odd`
        approximation using roots of polynomials orthogonal in
        measure |x|^alpha * epx(-x^2)

anisotropic_weights: list or array of weights (Int) length must be dimension or 2*dimension the first dimension weights must be positive see the manual for details

alpha, beta: Float64 alpha : the alpha parameter for Gegenbauer, Jacobi, Hermite and Laguerre rules beta : the beta parameter for Jacobi rules

custom_filename: string giving the path to the file with custom-tabulated rule

Tasmanian.makeLocalPolynomialGrid! — Method

makeLocalPolynomialGrid!(tsg::TasmanianSG; order=1, rule="localp", level_limits=Vector{Int32}(undef, 0))

creates a new sparse grid using a local polynomial rule discards any existing grid held by TSG

order: int (must be -1 or bigger) -1 indicates largest possible order 1 means linear, 2 means quadratic, etc. 0 means piece-wise constant, it has different hierarchy then the other orders, most notably the 1D rule triples the number of points per level (as opposed to double for the other cases)

rule: string (defines the 1-D rule that induces the grid) localp localp-zero semi-localp localp-boundary

Tasmanian.makeSequenceGrid! — Method

makeSequenceGrid!(tsg::TasmanianSG; type, rule, anisotropic_weights=Vector{Int32}(undef, 0), level_limits=Vector{Int32}(undef, 0))

creates a new sparse grid using a sequence rule discards any existing grid held by this class

type: string identifying the tensor selection strategy 'level' 'curved' 'hyperbolic' 'tensor' 'iptotal' 'ipcurved' 'iphyperbolic' 'iptensor' 'qptotal' 'qpcurved' 'qphyperbolic' 'qptensor'

rule: string (defines the 1-D rule that induces the grid) 'leja' 'rleja' 'rleja-shifted' 'max-lebesgue' 'min-lebesgue' 'min-delta'

anisotropic_weights: list or array of weights length must be tsg.dimensions or 2*tsg.dimensions the first tsg.dimensions weights must be positive see the manual for details

Tasmanian.makeWaveletGrid! — Method

makeWaveletGrid!(tsg::TasmanianSG; order=1, level_limits=Vector{Int32}(undef, 0))

creates a new sparse grid using a wavelet rule discards any existing grid held by tsg

order: Int (must be 1 or 3) only wavelets of order 1 and 3 are implemented

Tasmanian.mergeRefinement — Method

mergeRefinement(tsg::TasmanianSG)

combines the loaded and needed points into a single grid it also invalidates any currently loaded values, i.e., the grid cannot be used for internal integration or interpolation until loadNeededPoints() or setHierarchicalCoefficients() is called (even if those have been called before)

if getNumNeeded() == 0, this call will have no effect

Tasmanian.printStats — Method

printStats(tsg::TasmanianSG)

calls the library printStats() function, which displays basic information about this instance of the grid

Tasmanian.removePointsByHierarchicalCoefficient! — Function

removePointsByHierarchicalCoefficient!(tsg::TasmanianSG, tolerance, output = -1, scale_correction = [], NumKeep = -1)

removes any points in the grid with relative surplus that exceeds the tolerance or keeps the set number of points with largest surplus

tolerance: float (positive) the relative surplus tolerance, i.e., we keep only for points associated with surplus that exceeds the tolerance if NumKeep is positive, then tolerance is ignored

output: int (indicates the output to use) selects which output to consider accept -1 to indicate all outputs

scale_correction: vector or matrix if output = -1 and getNumOutputs() > 1, then using matrix with shape getNumOutputs() X getNumLoaded() with one weight per hierarchical coefficient if output > -1, then using a vector with one weight per point

NumKeep: int (positive or equal to -1) indicates the number of points to keep if set to -1 then tolerance is used as a cutoff if positive then the given number of points will be kept

Tasmanian.setAnisotropicRefinement! — Function

setAnisotropicRefinement!(tsg::TasmanianSG, type, min_growth, output, level_limits = Vector{Int32}(undef, 0))

estimates anisotropic coefficients from the current set of loaded points and updates the grid with the best points according to the estimate

type: string identifying the estimate to use (see the Manual) recommended: 'iptotal' 'ipcurved'

min_growth: int (positive) minimum number of new points to include in the new grid

output: int (indicates the output to use) selects which output to use for refinement sequence grids, using -1 indicates to use all outputs

level_limits: (if not empty) will be used to overwrite the currently set limits. The limits must be either empty or have size getNumDimensions(); if empty, the current set of limits will be used.

Tasmanian.setConformalTransformASIN! — Method

setConformalTransformASIN!(tsg::TasmanianSG, truncation)

sets conformal domain transform based on truncated Maclaurin series of arcsin()

truncation: a vector of non-negative integers indicating the truncation order in each direction 0 indicates no transform applied to this direction

Tasmanian.setDomainTransform! — Method

setDomainTransform!(tsg::TasmanianSG, Transform::VecOrMat)

sets the lower and upper bound for each dimension

Note: gauss-laguerre and gauss-hermite rules are defined on unbounded domain, in which case this sets the shift and scale parameters, consult the manual

Transform: a matrix of size dimension X 2 transform specifies the lower and upper bound of the domain in each direction.

       For gauss-laguerre and gauss-hermite grids, the
       transform gives the a and b parameters of the
       weights
        exp(-b (x - a))
        exp(-b (x - a)^2)

Tasmanian.setGPUID! — Method

setGPUID!(tsg::TasmanianSG, GPUID)

when using cuda on a machine with multiple GPUs, this helps set the GPU for this grid NOTE: each instance of the sparse grids class holds a separate instance of GPUID and different grids can be assigned to different GPUs (on multi-gpu system) GPUID can be changed at any time, however, this will cause some of the internal cache to be invalidated and it may lead to extraneous data movement

calling read or make***Grid will reset the selected GPU

defaults to 0

this doesn't do anything unless enableAcceleration is called using a "gpu-" acceleration type

Tasmanian.setHierarchicalCoefficients! — Method

setHierarchicalCoefficients!(tsg::TasmanianSG, coefficients)

Local polynomial, Wavelet, and Sequence grids construct approximation using hierarchical coefficients based on the loaded values. This function does the opposite, the hierarchical coefficients are loaded directly and the values are computed based on the coefficients. The coefficients can be computed, e.g., by solving least-squares or compressed sensing problem min || A c - f || where A is a matrix returned by evaluateHierarchicalFunctions() or evaluateSparseHierarchicalFunctions() for a set of points x; f are the values of the target function at the x points; and c is the vector with corresponding hierarchical coefficients.

If there is a pending refinement, i.e., getNumLoaded() != 0 and getNumNeeded() != 0, then the refinement is discarded (since it was computed based on the old and now obsolete values)

coefficients: a matrix with dimensions getNumOutputs() by getNumPoints() each column corresponds to the values of the coefficients at the corresponding point. The order and leading dimension must match the points obtained form getPoints(), the same order as the second dimension of evaluateHierarchicalFunctions(). This matrix must be of type Matrix{ComplexF64} when using a Fourier grid.

Tasmanian.setSurplusRefinement! — Method

setSurplusRefinement!(tsg::TasmanianSG, tol::Float64; output::Int=-1, refinement_type::AbstractString="", level_limits=Vector{Int32}(undef, 0), scale_correction = Vector{Float64}(undef, 0))

using hierarchical surplusses as an error indicator, the surplus refinement adds points to the grid to improve accuracy

when using sequence grids: this algorithm corresponds to the greedy Knapsack problem

when using local polynomial or wavelet grids, this call corresponds to local spatial refinement

tolerance: float (non-negative) the relative error tolerance, i.e., we refine only for points associated with surplus that exceeds the tolerance

output: int (indicates the output to use) selects which output to use for refinement sequence and local polynomial grids accept -1 to indicate all outputs

refinement_type: hierarchical and direction refinement strategy 'classic' 'parents' 'direction' 'fds' 'stable' applicable only for Local Polynomial and Wavelet grids

scalecorrection: matrix of non-negative numbers Instead of comparing the normalized surpluses to the tolerance, the scaled surplus will be used. The correction allows to manually guide the refinement process. The surplus of the j-th output of the i-th point will be scaled by scalecorrection[i][j]. scalecorrection.shape[0] must be equal to getNumLoaded() If empty, the scale is assumed 1.0 scalecorrection.shape[1] must be equal to the number of outputs used in the process, which is equal to getNumOutputs() for output == -1, or 1 if output > -1.

Tasmanian.updateFourierGrid! — Method

updateFourierGrid!(tsg::TasmanianSG, depth, type; anisotropic_weights=Vector{Int32}(undef, 0), level_limits=Vector{Int32}(undef, 0))

adds the points defined by depth, type and anisotropy to the existing grid

basically, the same as calling makeFourierGrid() with rule, of this grid and the new depth, type and anisotropic_weights then adding the resulting points to the current grid

inputs: see help(makeGlobalGrid)

Tasmanian.updateGlobalGrid! — Method

updateGlobalGrid!(tsg::TasmanianSG, depth, type; anisotropicweights=Vector{Int32}(undef, 0), levellimits=Vector{Int32}(undef, 0)) adds the points defined by depth, type and anisotropy to the existing grid

basically, the same as calling makeGlobalGrid with rule, alpha and beta of this grid and the new depth, type and anisotropic_weights then adding the resulting points to the current grid

inputs: see help(makeGlobalGrid)

Tasmanian.updateSequenceGrid! — Method

updateSequenceGrid!(tsg::TasmanianSG, depth, type; anisotropic_weights=Vector{Int32}(undef, 0), level_limits=Vector{Int32}(undef, 0))

adds the points defined by depth, type and anisotropy to the existing grid

basically, the same as calling makeSequenceGrid() with rule, of this grid and the new depth, type and anisotropic_weights then adding the resulting points to the current grid

inputs: see help(makeSequenceGrid)